split the encoder configuration between ticks count and scale
The former can be read from the controller configuration. It turns
out that, unlike Weg's documentation, it is *not* the number of ticks
per turn - in our controller, it is 1/4 of it. Maybe the ticks per
phase instead ?

I'm in the process of figuring out what it is by talking to Weg.
In the meantime, allow to explicitly configure the scale.