fix: separate current alarm from current fault
There are no status flags to indicate that there is an
alarm, we would have to read the whole fault state to know whether
there is one or not, so we need a separate "is there an alarm or not"
method.

Since there is very little information associated with the current
alarm, and the rest of the fault-related information is meaningless
to interpret it. I therefore decided to split the two completely.